Instantiate view controller via UIStoryboard
The storyboard contained the XML required to instantiate the button,
etc. on the view controller, while the unit tests initialized the view
controller via it's default initializer.

In order to get the button to be instantiated as well (and thus for
the tests to pass), the view controller needs to be initialized via the
storyboard. This can be accomplished using the
-instantiateViewControllerWithIdentifier method. That method requires
the view controller in the storyboard have an identifier, so give it
one.

In addition, move the source code for the view controller out of the
test target, and instead make it public. This probably isn't necessary
in order to get the test to pass, but it's my personal preference.